get_alerts - needs more refinements allowing the
Needs more refinements allowing the user to select from a list of
states. Needs to make sure uniqueness in the alerts table as well. The
README has better documentation and of course we have the model and
migration for the alerts table, which may need some additional tweaking
as well