get_alert_tweaks - can pass an option or
Can pass an option or none at all which triggers a prompt to select from
a list of states. It works and added some TODOs in the code for
something to come back to. Thinking soft deletion might not be a good
idea on the alerts table and just create a cron that clears out any past
alert based on the ends value